nuevos cambios implementados order, orderdetail, manejo de errores

+++ src/main/java/com/nttdata/dockerized/postgresql/service/CategoryServiceImpl.java
@@ -1,5 +1,7 @@
 package com.nttdata.dockerized.postgresql.service;
 
+import com.nttdata.dockerized.postgresql.exception.BadRequestException;
+import com.nttdata.dockerized.postgresql.exception.NotFoundException;
 import com.nttdata.dockerized.postgresql.model.entity.Category;
 import com.nttdata.dockerized.postgresql.repository.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,25 +22,34 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public Category findById(Long id) {
-        return categoryRepository.findById(id).orElse(null);
+        return categoryRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Categoria no encontrada","404"));
     }
 
     @Override
     public Category save(Category category) {
+
+        if (category.getName() == null || category.getName().isBlank()){
+            throw new BadRequestException("La categoria es obligatoria");
+        }
+
         return categoryRepository.save(category);
     }
 
     @Override
     public Category updateById(Long id, Category category) {
         return categoryRepository.findById(id)
-                .map(existing -> {
-                    existing.setName(category.getName());
-                    return categoryRepository.save(existing);
-                }).orElse(null);
+                .map(categoryExisting -> {
+                    categoryExisting.setName(category.getName());
+                    return categoryRepository.save(categoryExisting);
+                }).orElseThrow(() -> new NotFoundException("Categoria no encontrada para acttualizar", "404"));
     }
 
     @Override
     public void deleteById(Long id) {
+        if (categoryRepository.existsById(id)){
+            throw new NotFoundException("Categoria no encontrada para eliminar","404");
+        }
         categoryRepository.deleteById(id);
     }
 }
